180DC_Lodestar-OCR/academic_ocr/requirements.txt
google-genai>=1.0.0
typing_extensions>=4.11.0
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
